tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.initial.reset.2dstate.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.canvas.host.initial.reset.2dstate</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">

<h1>2d.canvas.host.initial.reset.2dstate</h1>

<script>

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.strokeStyle;
  ctx.strokeStyle = "#ff0000";
  canvas.width = 100;
  _assertSame(ctx.strokeStyle, default_val, "ctx.strokeStyle", "default_val");
}, "Resetting the canvas state resets strokeStyle");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.fillStyle;
  ctx.fillStyle = "#ff0000";
  canvas.width = 100;
  _assertSame(ctx.fillStyle, default_val, "ctx.fillStyle", "default_val");
}, "Resetting the canvas state resets fillStyle");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.globalAlpha;
  ctx.globalAlpha = 0.5;
  canvas.width = 100;
  _assertSame(ctx.globalAlpha, default_val, "ctx.globalAlpha", "default_val");
}, "Resetting the canvas state resets globalAlpha");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.lineWidth;
  ctx.lineWidth = 0.5;
  canvas.width = 100;
  _assertSame(ctx.lineWidth, default_val, "ctx.lineWidth", "default_val");
}, "Resetting the canvas state resets lineWidth");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.lineCap;
  ctx.lineCap = "round";
  canvas.width = 100;
  _assertSame(ctx.lineCap, default_val, "ctx.lineCap", "default_val");
}, "Resetting the canvas state resets lineCap");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.lineJoin;
  ctx.lineJoin = "round";
  canvas.width = 100;
  _assertSame(ctx.lineJoin, default_val, "ctx.lineJoin", "default_val");
}, "Resetting the canvas state resets lineJoin");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.miterLimit;
  ctx.miterLimit = 0.5;
  canvas.width = 100;
  _assertSame(ctx.miterLimit, default_val, "ctx.miterLimit", "default_val");
}, "Resetting the canvas state resets miterLimit");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.shadowOffsetX;
  ctx.shadowOffsetX = 5;
  canvas.width = 100;
  _assertSame(ctx.shadowOffsetX, default_val, "ctx.shadowOffsetX", "default_val");
}, "Resetting the canvas state resets shadowOffsetX");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.shadowOffsetY;
  ctx.shadowOffsetY = 5;
  canvas.width = 100;
  _assertSame(ctx.shadowOffsetY, default_val, "ctx.shadowOffsetY", "default_val");
}, "Resetting the canvas state resets shadowOffsetY");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.shadowBlur;
  ctx.shadowBlur = 5;
  canvas.width = 100;
  _assertSame(ctx.shadowBlur, default_val, "ctx.shadowBlur", "default_val");
}, "Resetting the canvas state resets shadowBlur");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.shadowColor;
  ctx.shadowColor = "#ff0000";
  canvas.width = 100;
  _assertSame(ctx.shadowColor, default_val, "ctx.shadowColor", "default_val");
}, "Resetting the canvas state resets shadowColor");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.globalCompositeOperation;
  ctx.globalCompositeOperation = "copy";
  canvas.width = 100;
  _assertSame(ctx.globalCompositeOperation, default_val, "ctx.globalCompositeOperation", "default_val");
}, "Resetting the canvas state resets globalCompositeOperation");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.font;
  ctx.font = "25px serif";
  canvas.width = 100;
  _assertSame(ctx.font, default_val, "ctx.font", "default_val");
}, "Resetting the canvas state resets font");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.textAlign;
  ctx.textAlign = "center";
  canvas.width = 100;
  _assertSame(ctx.textAlign, default_val, "ctx.textAlign", "default_val");
}, "Resetting the canvas state resets textAlign");

test(t => {
  const canvas = document.createElement('canvas');
  canvas.width = 100;
  canvas.height = 50;
  const ctx = canvas.getContext('2d');

  const default_val = ctx.textBaseline;
  ctx.textBaseline = "bottom";
  canvas.width = 100;
  _assertSame(ctx.textBaseline, default_val, "ctx.textBaseline", "default_val");
}, "Resetting the canvas state resets textBaseline");

</script>
